requirements.txt
openai>=1.46.0
pillow
pyautogui
pytesseract
python-dotenv
mss